Better Player status box

+++ src/DFRobotPlayerController.cpp
@@ -51,7 +51,7 @@ void DFRobotPlayerController::playSound(int track) {
 
 }
 
-void DFRobotPlayerController::playSound(int track, uint32_t durationMs) {
+void DFRobotPlayerController::playSound(int track, unsigned long durationMs) {
     // Call base class first for setting status and timing
     PlayerController::playSound(track, durationMs);
 
@@ -59,7 +59,7 @@ void DFRobotPlayerController::playSound(int track, uint32_t durationMs) {
 
 }
 
-void DFRobotPlayerController::playSound(int track, uint32_t durationMs, const char* trackName) {
+void DFRobotPlayerController::playSound(int track, unsigned long durationMs, const char* trackName) {
     // Call the base class implementation
     PlayerController::playSound(track, durationMs, trackName);
