+++ src/DFRobotPlayerController.h
@@ -20,8 +20,8 @@ public:
     DFRobotPlayerController(int rxPin, int txPin);
     void begin() override;
     void playSound(int track) override;
-    void playSound(int track, uint32_t durationMs) override;
-    void playSound(int track, uint32_t durationMs, const char* trackName) override;
+    void playSound(int track, unsigned long durationMs) override;
+    void playSound(int track, unsigned long durationMs, const char* trackName) override;
     void stopSound() override;
     void enableLoop() override;
     void disableLoop() override;
